+++ NEWS.md
@@ -1,3 +1,10 @@
+# opensimplex2 V0.0.3
+
+* Fixed undefined behaviour
+* Fixed CRAN warnings
+* Updated README
+* Added `vignette("opensimplex-on-fire")`
+
 # opensimplex2 V0.0.2
 
 * Initial Release. Implements `opensimplex_noise()` and `opensimplex_space()`